feat(theme-shadcn): add base subpath export for lightweight theme setup [#1143]
Extract configureThemeBase() into @vertz/theme-shadcn/base, returning only
{ theme, globals } without importing 38 style factories and 30+ component
factories. Landing page bundle drops from 293KB to 161KB raw (~74KB to ~43KB
gzipped).

- Add packages/theme-shadcn/src/base.ts with configureThemeBase()
- Refactor configureTheme() to delegate to configureThemeBase()
- Add ./base subpath export to package.json
- Update landing page to import from @vertz/theme-shadcn/base
- Update create-vertz-app template to use configureThemeBase by default
- Add @vertz/theme-shadcn to tree-shaking test (13.5% ratio)
