+++ packages/theme-shadcn/bunup.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'bunup';
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/configs.ts'],
+  entry: ['src/index.ts', 'src/configs.ts', 'src/base.ts'],
   dts: true,
 });
