+++ packages/theme-shadcn/src/index.ts
@@ -1,8 +1,10 @@
 export type {
   ResolvedTheme,
+  ResolvedThemeBase,
   ThemeComponents,
   ThemeConfig,
   ThemedPrimitives,
+  ThemeStyle,
   ThemeStyles,
 } from './configure';
 export { configureTheme } from './configure';
